+++ Tests/XcodeTests/UIKitProject/UIKitProject/CustomEntityMigrationPolicy.swift
@@ -1,0 +1,7 @@
+import Foundation
+import CoreData
+
+class CustomEntityMigrationPolicy: NSEntityMigrationPolicy {
+    override func begin(_ mapping: NSEntityMapping, with manager: NSMigrationManager) throws {
+    }
+}